latest sr_3.0 fixes, after kamailio 3.0 release.

* origin/sr_3.0: (58 commits)
  Backport of the changes (see previous commits).
  kex: documented missing functions
  kex: aliased avp_printf to pv_printf
  auth_diameter(k): remove redundant declaration of find_credentials, now in core
  - fix bug in encode_contact()
  - fix bug when RURI was changed before and newuri will be used
  - remove unused code
  modules_k/nathelper: removed garbage from documentation (credits to Klaus)
  modules_k/nathelper:  improved handle_ruri_alias() documentation
  modules/lcr and modules_k/nathelper: fixed compiler warnings
  modules_k/nathelper: handle_uri_alias() alias handling fix
  modules_k/usrloc: changed 'Kamailio' to 'SIP Router' in user agent string.
  tm: safer handling for local transactions and drop_replies!=0
  sctp: stats: don't increment ASSOC_SHUTDOWN on COMM_LOST
  sctp: SCTP_STATS_ASSOC_SHUTDOWN
  tm doc: local_ack_mode documentation
  tm: new param: local_ack_mode
  makefile: fix make bin &  basedir
  core: fix parsing for avps with the same name as a pv class
  tls: name the certificates based on $MAIN_NAME
  ...

Conflicts:
	modules/tm/README
	modules_k/nathelper/rtpproxy_stream.c
	pkg/debian/rules

+++ tcp_main.c
@@ -2817,8 +2817,9 @@ inline static int handle_tcp_child(struct tcp_child* tcp_c, int fd_i)
 	if (unlikely(tcp_c->unix_sock<=0)){
 		/* (we can't have a fd==0, 0 is never closed )*/
 		LOG(L_CRIT, "BUG: handle_tcp_child: fd %d for %d "
-				"(pid %d, ser no %d)\n", tcp_c->unix_sock,
-				(int)(tcp_c-&tcp_children[0]), tcp_c->pid, tcp_c->proc_no);
+				"(pid %ld, ser no %d)\n", tcp_c->unix_sock,
+				(int)(tcp_c-&tcp_children[0]), (long)tcp_c->pid,
+				 tcp_c->proc_no);
 		goto error;
 	}
 	/* read until sizeof(response)
@@ -2827,9 +2828,9 @@ inline static int handle_tcp_child(struct tcp_child* tcp_c, int fd_i)
 	if (unlikely(bytes<(int)sizeof(response))){
 		if (bytes==0){
 			/* EOF -> bad, child has died */
-			DBG("DBG: handle_tcp_child: dead tcp child %d (pid %d, no %d)"
+			DBG("DBG: handle_tcp_child: dead tcp child %d (pid %ld, no %d)"
 					" (shutting down?)\n", (int)(tcp_c-&tcp_children[0]), 
-					tcp_c->pid, tcp_c->proc_no );
+					(long)tcp_c->pid, tcp_c->proc_no );
 			/* don't listen on it any more */
 			io_watch_del(&io_h, tcp_c->unix_sock, fd_i, 0); 
 			goto error; /* eof. so no more io here, it's ok to return error */
@@ -2838,8 +2839,8 @@ inline static int handle_tcp_child(struct tcp_child* tcp_c, int fd_i)
 			 * e.g.: SIGIO_RT overflow mode or EPOLL ET */
 			if ((errno!=EAGAIN) && (errno!=EWOULDBLOCK)){
 				LOG(L_CRIT, "ERROR: handle_tcp_child: read from tcp child %ld "
-						" (pid %d, no %d) %s [%d]\n",
-						(long)(tcp_c-&tcp_children[0]), tcp_c->pid,
+						" (pid %ld, no %d) %s [%d]\n",
+						(long)(tcp_c-&tcp_children[0]), (long)tcp_c->pid,
 						tcp_c->proc_no, strerror(errno), errno );
 			}else{
 				bytes=0;
@@ -2864,8 +2865,8 @@ inline static int handle_tcp_child(struct tcp_child* tcp_c, int fd_i)
 	if (unlikely(tcpconn==0)){
 		/* should never happen */
 		LOG(L_CRIT, "BUG: handle_tcp_child: null tcpconn pointer received"
-				 " from tcp child %d (pid %d): %lx, %lx\n",
-				 	(int)(tcp_c-&tcp_children[0]), tcp_c->pid,
+				 " from tcp child %d (pid %ld): %lx, %lx\n",
+				 	(int)(tcp_c-&tcp_children[0]), (long)tcp_c->pid,
 					response[0], response[1]) ;
 		goto end;
 	}
@@ -3337,9 +3338,9 @@ inline static int send2child(struct tcp_connection* tcpconn)
 				" connection passed to the least busy one (%d)\n",
 				min_busy);
 	}
-	DBG("send2child: to tcp child %d %d(%d), %p\n", idx, 
+	DBG("send2child: to tcp child %d %d(%ld), %p\n", idx, 
 					tcp_children[idx].proc_no,
-					tcp_children[idx].pid, tcpconn);
+					(long)tcp_children[idx].pid, tcpconn);
 	/* first make sure this child doesn't have pending request for
 	 * tcp_main (to avoid a possible deadlock: e.g. child wants to
 	 * send a release command, but the master fills its socket buffer
